Add getMeasureStart function
