+++ src/main/actions/track.ts
@@ -1,4 +1,3 @@
-import { maxBy } from "lodash"
 import { AnyChannelEvent, SetTempoEvent } from "midifile-ts"
 import {
   panMidiEvent,
@@ -7,12 +6,8 @@ import {
   timeSignatureMidiEvent,
   volumeMidiEvent,
 } from "../../common/midi/MidiEvent"
-import {
-  isNoteEvent,
-  isTimeSignatureEvent,
-  NoteEvent,
-  TrackEventOf,
-} from "../../common/track"
+import { getMeasureStart } from "../../common/song/selector"
+import { isNoteEvent, NoteEvent, TrackEventOf } from "../../common/track"
 import RootStore from "../stores/RootStore"
 import { pushHistory } from "./history"
 import {
@@ -336,38 +331,16 @@ export const addTimeSignature =
   (tick: number, numerator: number, denominator: number) => {
     const { song } = rootStore
 
-    if (song.conductorTrack === undefined) {
+    const measureStart = getMeasureStart(song, tick)
+
+    const timeSignatureTick = measureStart?.tick ?? 0
+
+    // prevent duplication
+    if (
+      measureStart !== null &&
+      measureStart.timeSignature.tick === measureStart.tick
+    ) {
       return
-    }
-
-    // get the nearest time signature
-    const timeSignature = maxBy(
-      song.conductorTrack.events
-        .filter(isTimeSignatureEvent)
-        .slice()
-        .filter((e) => e.tick <= tick),
-      (e) => e.tick
-    )
-
-    let timeSignatureTick: number
-
-    if (timeSignature === undefined) {
-      timeSignatureTick = 0
-    } else {
-      // calculate the nearest measure beginning
-      const ticksPerMeasure =
-        ((song.timebase * 4) / timeSignature.denominator) *
-        timeSignature.numerator
-      const numberOfMeasures = Math.floor(
-        (tick - timeSignature.tick) / ticksPerMeasure
-      )
-      timeSignatureTick =
-        timeSignature.tick + numberOfMeasures * ticksPerMeasure
-
-      // prevent duplication
-      if (timeSignature.tick === timeSignatureTick) {
-        return
-      }
     }
 
     pushHistory(rootStore)()
